Add basic driver functionality
Adds most of the core functionality needed to configure the BME68x and retrieve data from it. TODO comments indicate where remaining functionality needs to be added or modified. The primary remaining function to implement is a microsecond delay callback, which is used to insert the necessary wait periods for sensor measurements to take place.