Add outline of test code for BME68x driver
Adds a basic example of how to use the driver. This is considered an outline instead of a fully-functioning example, as the data retrieved is currently all zeros rather than actual sensor readings. Upon resolution of the issue, this code should work to showcase driver initialization, configuration, and retrieval of data from the sensor.